Extract function_template.h and .cc
