+++ function_template.go
@@ -5,7 +5,7 @@
 package v8go
 
 // #include <stdlib.h>
-// #include "v8go.h"
+// #include "function_template.h"
 import "C"
 import (
 	"runtime"
